+++ Assets/BeauRoutine/Routine/Sequence.cs
@@ -71,6 +71,15 @@ namespace BeauRoutine
         }
 
         /// <summary>
+        /// Executes the routines in parallel.
+        /// </summary>
+        public Sequence Combine(params IEnumerator[] inNexts)
+        {
+            AddNode(NodeType.Routine, Routine.Combine(inNexts));
+            return this;
+        }
+
+        /// <summary>
         /// Waits for the given number of seconds.
         /// </summary>
         public Sequence Wait(float inSeconds)
@@ -128,12 +137,12 @@ namespace BeauRoutine
 
         #region IRoutineEnumerator
 
-        public object Current
+        object IEnumerator.Current
         {
             get { return m_Current; }
         }
 
-        public bool MoveNext()
+        bool IEnumerator.MoveNext()
         {
             m_Current = null;
             while(true)
@@ -161,6 +170,11 @@ namespace BeauRoutine
             }
         }
 
+        void IEnumerator.Reset()
+        {
+            throw new NotSupportedException();
+        }
+
         public void Dispose()
         {
             for(int i = m_Index + 1; i < m_Nodes.Count; ++i)
@@ -172,11 +186,6 @@ namespace BeauRoutine
                         break;
                 }
             }
-        }
-
-        public void Reset()
-        {
-            throw new NotSupportedException();
         }
 
         #endregion
