+++ Assets/BeauRoutine/Tween/Shortcuts/TweenShortcuts.Canvas.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 3e66e48316be08f4ea51cdd5e73e45f0
+timeCreated: 1525809061
+licenseType: Pro
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
