+++ Assets/BeauRoutine/Tween/Shortcuts/TweenShortcuts.Renderer.cs
@@ -18,376 +18,6 @@ namespace BeauRoutine
     /// </summary>
     static public partial class TweenShortcuts
     {
-        #region CanvasGroup
-
-        private sealed class TweenData_CanvasGroup_Alpha : ITweenData
-        {
-            private CanvasGroup m_Renderer;
-            private float m_Target;
-
-            private float m_Start;
-            private float m_Delta;
-
-            public TweenData_CanvasGroup_Alpha(CanvasGroup inRenderer, float inTarget)
-            {
-                m_Renderer = inRenderer;
-                m_Target = inTarget;
-            }
-
-            public void OnTweenStart()
-            {
-                m_Start = m_Renderer.alpha;
-                m_Delta = m_Target - m_Start;
-            }
-
-            public void ApplyTween(float inPercent)
-            {
-                m_Renderer.alpha = m_Start + m_Delta * inPercent;
-            }
-
-            public void OnTweenEnd() { }
-
-            public override string ToString()
-            {
-                return "CanvasGroup: Alpha";
-            }
-        }
-
-        /// <summary>
-        /// Fades the CanvasGroup to another alpha over time.
-        /// </summary>
-        static public Tween FadeTo(this CanvasGroup inRenderer, float inAlpha, float inTime)
-        {
-            return Tween.Create(new TweenData_CanvasGroup_Alpha(inRenderer, inAlpha), inTime);
-        }
-
-        /// <summary>
-        /// Fades the CanvasGroup to another alpha over time.
-        /// </summary>
-        static public Tween FadeTo(this CanvasGroup inRenderer, float inAlpha, TweenSettings inSettings)
-        {
-            return Tween.Create(new TweenData_CanvasGroup_Alpha(inRenderer, inAlpha), inSettings);
-        }
-
-        #endregion
-
-        #region Graphic
-
-        private sealed class TweenData_Graphic_Alpha : ITweenData
-        {
-            private Graphic m_Renderer;
-            private float m_Target;
-
-            private float m_Start;
-            private float m_Delta;
-
-            public TweenData_Graphic_Alpha(Graphic inRenderer, float inTarget)
-            {
-                m_Renderer = inRenderer;
-                m_Target = inTarget;
-            }
-
-            public void OnTweenStart()
-            {
-                m_Start = m_Renderer.color.a;
-                m_Delta = m_Target - m_Start;
-            }
-
-            public void OnTweenEnd() { }
-
-            public void ApplyTween(float inPercent)
-            {
-                Color final = m_Renderer.color;
-                final.a = m_Start + m_Delta * inPercent;
-                m_Renderer.color = final;
-            }
-
-            public override string ToString()
-            {
-                return "Graphic: Alpha";
-            }
-        }
-
-        private sealed class TweenData_Graphic_Color : ITweenData
-        {
-            private Graphic m_Renderer;
-            private Color m_Target;
-            private ColorUpdate m_Update;
-
-            private Color m_Start;
-
-            public TweenData_Graphic_Color(Graphic inRenderer, Color inTarget, ColorUpdate inUpdate)
-            {
-                m_Renderer = inRenderer;
-                m_Target = inTarget;
-                m_Update = inUpdate;
-            }
-
-            public void OnTweenStart()
-            {
-                m_Start = m_Renderer.color;
-            }
-
-            public void OnTweenEnd() { }
-
-            public void ApplyTween(float inPercent)
-            {
-                Color final = UnityEngine.Color.LerpUnclamped(m_Start, m_Target, inPercent);
-                if (m_Update == ColorUpdate.PreserveAlpha)
-                    final.a = m_Renderer.color.a;
-                m_Renderer.color = final;
-            }
-
-            public override string ToString()
-            {
-                return "Graphic: Color";
-            }
-        }
-
-        private sealed class TweenData_Graphic_Gradient : ITweenData
-        {
-            private Graphic m_Renderer;
-            private Gradient m_Gradient;
-            private ColorUpdate m_Update;
-
-            public TweenData_Graphic_Gradient(Graphic inRenderer, Gradient inTarget, ColorUpdate inUpdate)
-            {
-                m_Renderer = inRenderer;
-                m_Gradient = inTarget;
-                m_Update = inUpdate;
-            }
-
-            public void OnTweenStart() { }
-            public void OnTweenEnd() { }
-
-            public void ApplyTween(float inPercent)
-            {
-                Color final = m_Gradient.Evaluate(inPercent);
-                if (m_Update == ColorUpdate.PreserveAlpha)
-                    final.a = m_Renderer.color.a;
-                m_Renderer.color = final;
-            }
-
-            public override string ToString()
-            {
-                return "Graphic: Gradient";
-            }
-        }
-
-        /// <summary>
-        /// Fades the Graphic to another alpha over time.
-        /// </summary>
-        static public Tween FadeTo(this Graphic inRenderer, float inAlpha, float inTime)
-        {
-            return Tween.Create(new TweenData_Graphic_Alpha(inRenderer, inAlpha), inTime);
-        }
-
-        /// <summary>
-        /// Fades the Graphic to another alpha over time.
-        /// </summary>
-        static public Tween FadeTo(this Graphic inRenderer, float inAlpha, TweenSettings inSettings)
-        {
-            return Tween.Create(new TweenData_Graphic_Alpha(inRenderer, inAlpha), inSettings);
-        }
-
-        /// <summary>
-        /// Fades the Graphic to another color over time.
-        /// </summary>
-        static public Tween ColorTo(this Graphic inRenderer, Color inTarget, float inTime, ColorUpdate inUpdate = ColorUpdate.PreserveAlpha)
-        {
-            return Tween.Create(new TweenData_Graphic_Color(inRenderer, inTarget, inUpdate), inTime);
-        }
-
-        /// <summary>
-        /// Fades the Graphic to another color over time.
-        /// </summary>
-        static public Tween ColorTo(this Graphic inRenderer, Color inTarget, TweenSettings inSettings, ColorUpdate inUpdate = ColorUpdate.PreserveAlpha)
-        {
-            return Tween.Create(new TweenData_Graphic_Color(inRenderer, inTarget, inUpdate), inSettings);
-        }
-
-        /// <summary>
-        /// Applies a gradient of colors to the Graphic over time.
-        /// </summary>
-        static public Tween Gradient(this Graphic inRenderer, Gradient inGradient, float inTime, ColorUpdate inUpdate = ColorUpdate.PreserveAlpha)
-        {
-            return Tween.Create(new TweenData_Graphic_Gradient(inRenderer, inGradient, inUpdate), inTime);
-        }
-
-        /// <summary>
-        /// Applies a gradient of colors to the Graphic over time.
-        /// </summary>
-        static public Tween Gradient(this Graphic inRenderer, Gradient inGradient, TweenSettings inSettings, ColorUpdate inUpdate = ColorUpdate.PreserveAlpha)
-        {
-            return Tween.Create(new TweenData_Graphic_Gradient(inRenderer, inGradient, inUpdate), inSettings);
-        }
-
-        #endregion
-
-        #region Image
-
-        private sealed class TweenData_Image_FillAmount : ITweenData
-        {
-            private Image m_Renderer;
-            private float m_Target;
-
-            private float m_Start;
-            private float m_Delta;
-
-            public TweenData_Image_FillAmount(Image inRenderer, float inTarget)
-            {
-                m_Renderer = inRenderer;
-                m_Target = inTarget;
-            }
-
-            public void OnTweenStart()
-            {
-                m_Start = m_Renderer.fillAmount;
-                m_Delta = m_Target - m_Start;
-            }
-
-            public void OnTweenEnd() { }
-
-            public void ApplyTween(float inPercent)
-            {
-                m_Renderer.fillAmount = m_Start + m_Delta * inPercent;
-            }
-
-            public override string ToString()
-            {
-                return "Image: FillAmount";
-            }
-        }
-
-        /// <summary>
-        /// Shifts the Image's fillAmount over time.
-        /// </summary>
-        static public Tween FillTo(this Image inRenderer, float inFillAmount, float inTime)
-        {
-            return Tween.Create(new TweenData_Image_FillAmount(inRenderer, inFillAmount), inTime);
-        }
-
-        /// <summary>
-        /// Shifts the Image's fillAmount over time.
-        /// </summary>
-        static public Tween FillTo(this Image inRenderer, float inFillAmount, TweenSettings inSettings)
-        {
-            return Tween.Create(new TweenData_Image_FillAmount(inRenderer, inFillAmount), inSettings);
-        }
-
-        #endregion
-
-        #region RawImage
-
-        private sealed class TweenData_RawImage_UVRect : ITweenData
-        {
-            private RawImage m_Renderer;
-            private Rect m_Target;
-            private Axis m_Axis;
-
-            private Rect m_Start;
-            private Vector4 m_Delta;
-
-            public TweenData_RawImage_UVRect(RawImage inRenderer, Rect inTarget, Axis inAxis)
-            {
-                m_Renderer = inRenderer;
-                m_Target = inTarget;
-                m_Axis = inAxis;
-            }
-
-            public void OnTweenStart()
-            {
-                m_Start = m_Renderer.uvRect;
-                m_Delta = new Vector4(m_Target.x - m_Start.x, m_Target.y - m_Start.y, m_Target.width - m_Start.width, m_Target.height - m_Start.height);
-            }
-
-            public void OnTweenEnd() { }
-
-            public void ApplyTween(float inPercent)
-            {
-                Rect final = new Rect(m_Start.x + m_Delta.x * inPercent,
-                    m_Start.y + m_Delta.y * inPercent,
-                    m_Start.width + m_Delta.z * inPercent,
-                    m_Start.height + m_Delta.w * inPercent);
-
-                Rect rendererRect = m_Renderer.uvRect;
-
-                if ((m_Axis & Axis.X) == 0)
-                {
-                    final.xMin = rendererRect.xMin;
-                    final.xMax = rendererRect.xMax;
-                }
-                if ((m_Axis & Axis.Y) == 0)
-                {
-                    final.yMin = rendererRect.yMin;
-                    final.yMax = rendererRect.yMax;
-                }
-
-                m_Renderer.uvRect = final;
-            }
-
-            public override string ToString()
-            {
-                return "RawImage: UVRect";
-            }
-        }
-
-        /// <summary>
-        /// Shifts the RawImage's uvRect over time.
-        /// </summary>
-        static public Tween UVRectShift(this RawImage inRenderer, Vector2 inShift, float inTime, Axis inAxis = Axis.XY)
-        {
-            Rect shiftedRect = inRenderer.uvRect;
-            shiftedRect.x += inShift.x;
-            shiftedRect.y += inShift.y;
-            return Tween.Create(new TweenData_RawImage_UVRect(inRenderer, shiftedRect, inAxis), inTime);
-        }
-
-        /// <summary>
-        /// Shifts the RawImage's uvRect over time.
-        /// </summary>
-        static public Tween UVRectShift(this RawImage inRenderer, float inShift, float inTime, Axis inAxis)
-        {
-            return UVRectShift(inRenderer, new Vector2(inShift, inShift), inTime, inAxis);
-        }
-
-        /// <summary>
-        /// Shifts the RawImage's uvRect over time.
-        /// </summary>
-        static public Tween UVRectShift(this RawImage inRenderer, Vector2 inShift, TweenSettings inSettings, Axis inAxis = Axis.XY)
-        {
-            Rect shiftedRect = inRenderer.uvRect;
-            shiftedRect.x += inShift.x;
-            shiftedRect.y += inShift.y;
-            return Tween.Create(new TweenData_RawImage_UVRect(inRenderer, shiftedRect, inAxis), inSettings);
-        }
-
-        /// <summary>
-        /// Shifts the RawImage's uvRect over time.
-        /// </summary>
-        static public Tween UVRectShift(this RawImage inRenderer, float inShift, TweenSettings inSettings, Axis inAxis)
-        {
-            return UVRectShift(inRenderer, new Vector2(inShift, inShift), inSettings, inAxis);
-        }
-
-        /// <summary>
-        /// Shifts the RawImage's uvRect over time.
-        /// </summary>
-        static public Tween UVRectTo(this RawImage inRenderer, Rect inTarget, float inTime, Axis inAxis = Axis.XY)
-        {
-            return Tween.Create(new TweenData_RawImage_UVRect(inRenderer, inTarget, inAxis), inTime);
-        }
-
-        /// <summary>
-        /// Shifts the RawImage's uvRect over time.
-        /// </summary>
-        static public Tween UVRectTo(this RawImage inRenderer, Rect inTarget, TweenSettings inSettings, Axis inAxis = Axis.XY)
-        {
-            return Tween.Create(new TweenData_RawImage_UVRect(inRenderer, inTarget, inAxis), inSettings);
-        }
-
-        #endregion
-
         #region SpriteRenderer
 
         private sealed class TweenData_SpriteRenderer_Alpha : ITweenData
@@ -694,153 +324,57 @@ namespace BeauRoutine
 
         #endregion
 
-        #region CanvasRenderer
+        #region SkinnedMeshRenderer
 
-        private sealed class TweenData_CanvasRenderer_Alpha : ITweenData
+        private sealed class TweenData_SkinnedMeshRenderer_BlendWeight : ITweenData
         {
-            private CanvasRenderer m_Renderer;
+            private SkinnedMeshRenderer m_Renderer;
+            private int m_Index;
             private float m_Target;
 
             private float m_Start;
             private float m_Delta;
 
-            public TweenData_CanvasRenderer_Alpha(CanvasRenderer inRenderer, float inTarget)
+            public TweenData_SkinnedMeshRenderer_BlendWeight(SkinnedMeshRenderer inRenderer, int inIndex, float inTarget)
             {
                 m_Renderer = inRenderer;
+                m_Index = inIndex;
                 m_Target = inTarget;
             }
 
             public void OnTweenStart()
             {
-                m_Start = m_Renderer.GetAlpha();
+                m_Start = m_Renderer.GetBlendShapeWeight(m_Index);
                 m_Delta = m_Target - m_Start;
             }
 
-            public void OnTweenEnd() { }
-
             public void ApplyTween(float inPercent)
             {
-                m_Renderer.SetAlpha(m_Start + m_Delta * inPercent);
-            }
-
-            public override string ToString()
-            {
-                return "CanvasRenderer: Alpha";
-            }
-        }
-
-        private sealed class TweenData_CanvasRenderer_Color : ITweenData
-        {
-            private CanvasRenderer m_Renderer;
-            private Color m_Target;
-            private ColorUpdate m_Update;
-
-            private Color m_Start;
-
-            public TweenData_CanvasRenderer_Color(CanvasRenderer inRenderer, Color inTarget, ColorUpdate inUpdate)
-            {
-                m_Renderer = inRenderer;
-                m_Target = inTarget;
-                m_Update = inUpdate;
-            }
-
-            public void OnTweenStart()
-            {
-                m_Start = m_Renderer.GetColor();
+                m_Renderer.SetBlendShapeWeight(m_Index, m_Start + m_Delta * inPercent);
             }
 
             public void OnTweenEnd() { }
 
-            public void ApplyTween(float inPercent)
-            {
-                Color final = UnityEngine.Color.LerpUnclamped(m_Start, m_Target, inPercent);
-                if (m_Update == ColorUpdate.PreserveAlpha)
-                    final.a = m_Renderer.GetAlpha();
-                m_Renderer.SetColor(final);
-            }
-
             public override string ToString()
             {
-                return "CanvasRenderer: Color";
-            }
-        }
-
-        private sealed class TweenData_CanvasRenderer_Gradient : ITweenData
-        {
-            private CanvasRenderer m_Renderer;
-            private Gradient m_Gradient;
-            private ColorUpdate m_Update;
-
-            public TweenData_CanvasRenderer_Gradient(CanvasRenderer inRenderer, Gradient inTarget, ColorUpdate inUpdate)
-            {
-                m_Renderer = inRenderer;
-                m_Gradient = inTarget;
-                m_Update = inUpdate;
-            }
-
-            public void OnTweenStart() { }
-            public void OnTweenEnd() { }
-
-            public void ApplyTween(float inPercent)
-            {
-                Color final = m_Gradient.Evaluate(inPercent);
-                if (m_Update == ColorUpdate.PreserveAlpha)
-                    final.a = m_Renderer.GetAlpha();
-                m_Renderer.SetColor(final);
-            }
-
-            public override string ToString()
-            {
-                return "CanvasRenderer: Gradient";
+                return "SkinnedMeshRenderer: Blend Shape Weight";
             }
         }
 
         /// <summary>
-        /// Fades the Graphic to another alpha over time.
+        /// Tweens the weight for a blend shape on a SkinnedMeshRenderer to another value over time.
         /// </summary>
-        static public Tween FadeTo(this CanvasRenderer inRenderer, float inAlpha, float inTime)
+        static public Tween BlendShapeTo(this SkinnedMeshRenderer inRenderer, int inIndex, float inTarget, float inTime)
         {
-            return Tween.Create(new TweenData_CanvasRenderer_Alpha(inRenderer, inAlpha), inTime);
+            return Tween.Create(new TweenData_SkinnedMeshRenderer_BlendWeight(inRenderer, inIndex, inTarget), inTime);
         }
 
         /// <summary>
-        /// Fades the Graphic to another alpha over time.
+        /// Tweens the weight for a blend shape on a SkinnedMeshRenderer to another value over time.
         /// </summary>
-        static public Tween FadeTo(this CanvasRenderer inRenderer, float inAlpha, TweenSettings inSettings)
+        static public Tween BlendShapeTo(this SkinnedMeshRenderer inRenderer, int inIndex, float inTarget, TweenSettings inSettings)
         {
-            return Tween.Create(new TweenData_CanvasRenderer_Alpha(inRenderer, inAlpha), inSettings);
-        }
-
-        /// <summary>
-        /// Fades the Graphic to another color over time.
-        /// </summary>
-        static public Tween ColorTo(this CanvasRenderer inRenderer, Color inTarget, float inTime, ColorUpdate inUpdate = ColorUpdate.PreserveAlpha)
-        {
-            return Tween.Create(new TweenData_CanvasRenderer_Color(inRenderer, inTarget, inUpdate), inTime);
-        }
-
-        /// <summary>
-        /// Fades the Graphic to another color over time.
-        /// </summary>
-        static public Tween ColorTo(this CanvasRenderer inRenderer, Color inTarget, TweenSettings inSettings, ColorUpdate inUpdate = ColorUpdate.PreserveAlpha)
-        {
-            return Tween.Create(new TweenData_CanvasRenderer_Color(inRenderer, inTarget, inUpdate), inSettings);
-        }
-
-        /// <summary>
-        /// Applies a gradient of colors to the Graphic over time.
-        /// </summary>
-        static public Tween Gradient(this CanvasRenderer inRenderer, Gradient inGradient, float inTime, ColorUpdate inUpdate = ColorUpdate.PreserveAlpha)
-        {
-            return Tween.Create(new TweenData_CanvasRenderer_Gradient(inRenderer, inGradient, inUpdate), inTime);
-        }
-
-        /// <summary>
-        /// Applies a gradient of colors to the Graphic over time.
-        /// </summary>
-        static public Tween Gradient(this CanvasRenderer inRenderer, Gradient inGradient, TweenSettings inSettings, ColorUpdate inUpdate = ColorUpdate.PreserveAlpha)
-        {
-            return Tween.Create(new TweenData_CanvasRenderer_Gradient(inRenderer, inGradient, inUpdate), inSettings);
+            return Tween.Create(new TweenData_SkinnedMeshRenderer_BlendWeight(inRenderer, inIndex, inTarget), inSettings);
         }
 
         #endregion
