+++ Assets/Tests/8 - ThinkUpdate/Example8_Script.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections;
 using BeauRoutine;
+using BeauRoutine.Splines;
 using UnityEngine;
 using UnityEngine.UI;
 
@@ -7,6 +8,8 @@ namespace BeauRoutine.Examples
 {
     public class Example8_Script : MonoBehaviour
     {
+        public VertexSpline Spline;
+
         private IEnumerator Start()
         {
             yield return null;
@@ -14,12 +17,16 @@ namespace BeauRoutine.Examples
             // Routine.Start(this, Executing("CustomUpdate: ")).SetPhase(RoutinePhase.CustomUpdate);
 
             // PerSecondRoutine.Start(this, Executing("PerSecond: "), 5);
-            Routine.Start(this, Routine.PerSecond(Executing("PerSecond: "), 5));
+            // Routine.Start(this, Routine.PerSecond(Executing("PerSecond: "), 5));
             //Routine.Start(this, Routine.PerSecond(started, 5f));
 
             Routine.Start(this, SquashStretch());
 
-            Routine.Start(this, Camera.main.BackgroundColorTo(Color.black, 0.5f).YoyoLoop());
+            // Routine.Start(this, Camera.main.BackgroundColorTo(Color.black, 0.5f).YoyoLoop());
+
+            transform.SetPosition(transform.position + Random.onUnitSphere * 10);
+
+            Routine.Start(this, transform.MoveAlong(Spline, 3, Axis.XYZ, Space.Self, SplinePositioning.Absolute).Loop().Randomize());
         }
 
         private IEnumerator Executing(string inPrefix)
@@ -37,7 +44,7 @@ namespace BeauRoutine.Examples
 
         private IEnumerator SquashStretch()
         {
-            yield return transform.SquashStretchTo(1.2f, 2f, Axis.Y, Axis.X).YoyoLoop().Wave(Wave.Function.CosFade, 5f);
+            yield return transform.SquashStretchTo(1.2f, 2f, Axis.Y, Axis.X).YoyoLoop().Wave(Wave.Function.CosFade, 5f).Randomize();
         }
     }
 }
